Move format-diff.js to src/util/
